Rewrite wheel generation/init now that I understand it.
This means `compute_wheel_elem` is now O(1) rather than O(low - p * p),
so that's nice.

It also dramatically simplifies the wheel generator, now that I'm not
doing random stuff with linear equations etc. Includes a description of
what's going on so hopefully I'll be able to reconstruct the thinking in
future?